+++ data/js/platforms/hitbox.js
@@ -15,13 +15,14 @@ websites.hitbox = {
 			}
 			return obj;
 		},
-	"isValidResponse":
+	"checkResponseValidity":
 		function(data){
 			if(data.error == "live"){
 				return "error";
 			}
 			if(data.error == true){
-				return "error";
+				let error_msg = (data.hasOwnProperty("error_msg") == true)? data.error_msg : "error";
+				return error_msg;
 			}
 			return "success";
 		},
@@ -30,7 +31,7 @@ websites.hitbox = {
 			let streamData = currentLiveStatus;
 			if(data.hasOwnProperty("livestream") == false){
 				if(data.error_msg="no_media_found"){
-					streamData.online = false;
+					streamData.liveStatus.API_Status = false;
 				}
 				streamData.streamName = id;
 				return null;
@@ -65,11 +66,11 @@ websites.hitbox = {
 				}
 				streamData.streamCurrentViewers = parseInt(data["media_views"]);
 				
-				streamData.online = (data["media_is_live"] == "1")? true : false;
+				streamData.liveStatus.API_Status = (data["media_is_live"] == "1")? true : false;
 				if(typeof data.channel["twitter_account"] == "string" && data.channel["twitter_account"] != "" && typeof data.channel["twitter_account"] == "string" && data.channel["twitter_enabled"] == "1"){
 					streamData.twitterID = data.channel["twitter_account"];
 				}
-				return streamData.online;
+				return streamData.liveStatus.API_Status;
 			} else {
 				return null;
 			}