+++ data/js/platforms/twitch.js
@@ -23,12 +23,18 @@ websites.twitch = {
 			}
 			return obj;
 		},
-	"isValidResponse":
+	"checkResponseValidity":
 		function(data){
-			if(data.error == "Not Found"){
-				return "error";
+			if(data.hasOwnProperty("error")){
+				if(typeof data.message == "string"){
+					return data.message;
+				} else if(typeof data.error == "string"){
+					return data.error;
+				} else {
+					return "error";
+				}
 			} else {
-				return "";
+				return "success";
 			}
 		},
 	"checkLiveStatus":
@@ -36,6 +42,7 @@ websites.twitch = {
 			let streamData = currentLiveStatus;
 			if(data.hasOwnProperty("stream")){
 				data = data["stream"];
+				streamData.liveStatus.API_Status = (data != null);
 				if(data != null){
 					streamData.streamName = data["channel"]["display_name"];
 					streamData.streamStatus = (data["channel"]["status"] != null)? data["channel"]["status"] : "";
@@ -48,14 +55,12 @@ websites.twitch = {
 					}
 					streamData.streamCurrentViewers = parseInt(data["viewers"]);
 					
-					streamData.online = true;
-					return streamData.online;
+					return streamData.liveStatus.API_Status;
 				} else {
 					if(streamData.streamName == ""){
 						streamData.streamName = id;
 					}
-					streamData.online = false;
-					return streamData.online;
+					return streamData.liveStatus.API_Status;
 				}
 			} else {
 				return null;